+++ libticables/branches/ticables-2/src/ticables.h
@@ -148,8 +148,8 @@ struct _CableFncts
 	int (*reset)	(CableHandle *);
 	int (*probe)	(CableHandle *);
 
-	int (*send)		(CableHandle *, uint8_t *, uint16_t);
-	int (*recv)		(CableHandle *, uint8_t *, uint16_t);
+	int (*send)		(CableHandle *, uint8_t *, uint32_t);
+	int (*recv)		(CableHandle *, uint8_t *, uint32_t);
 	int (*check)	(CableHandle *, int *);
 
 	int (*set_d0)	(CableHandle *, int);
@@ -232,8 +232,8 @@ struct _CableHandle
 	TIEXPORT int TICALL ticables_cable_reset(CableHandle* handle);
 	TIEXPORT int TICALL ticables_cable_probe(CableHandle*, unsigned int* result);
 
-	TIEXPORT int TICALL ticables_cable_send(CableHandle*, uint8_t *data, uint16_t len);
-	TIEXPORT int TICALL ticables_cable_recv(CableHandle*, uint8_t *data, uint16_t len);
+	TIEXPORT int TICALL ticables_cable_send(CableHandle*, uint8_t *data, uint32_t len);
+	TIEXPORT int TICALL ticables_cable_recv(CableHandle*, uint8_t *data, uint32_t len);
 
 	TIEXPORT int TICALL ticables_cable_check(CableHandle*, CableStatus*);
